switch to permissionless roles at bottom of list
Major changes to the function of ColorMe, to elimnate role hierarchy
conflicts. All colorme roles will now be permissionless roles at the
bottom of the list. This is a vast improvement to security. However,
note that other roles used by your members (for example, admin or mod
roles) must be set to the default color in order for the ColorMe color
to show. There is an added command, [p]colorme clean, which will attempt
to remove all permissions from ColorMe roles, to aid backwards
compatibility. As always, use caution with this command and manually
verify the roles that it lists before confirming!