Manage server is manage_guild
I am really hoping to use this cog as part of V3. I used it extensively in V2. This error came up when attempting to install it but it threw an error. If this can be fixed (i believe this is the correct change for the error) then it would be much appreciated.