[ReactPoll] Fix close_poll + Black formatting
Resolves #102 (or at least seems to in my testing)
The close_poll func was choking at multiple points and would not return a traceback for any of the bad conditions. When this occurred, the loop would break and it would stop checking for invalid polls.
This will close all invalid polls/polls that should have had an end message, this will result in a lot of messages sent out in some cases. Clearing out old polls can take a little while depending on how many need to be closed. My test bot with 10 invalid polls took about 10 sec on startup to verify and close all old polls.
The loop checker has been set to 5 seconds instead of 1 second, so poll endings may now be a little loose on the timing.